FAT-28387: Move delete authority scenarios to DataImportCriticalPathApiTest to match TestRails group
